internal PlatformOptimizedCancellationException
