+++ compose/foundation/foundation/src/commonMain/kotlin/androidx/compose/foundation/Expect.kt
@@ -18,6 +18,8 @@
 
 package androidx.compose.foundation
 
+import kotlinx.coroutines.CancellationException
+
 expect class AtomicReference<V>(value: V) {
     fun get(): V
     fun set(value: V)
@@ -30,3 +32,11 @@ expect class AtomicLong(value: Long) {
     fun set(value: Long)
     fun getAndIncrement(): Long
 }
+
+/**
+ * Represents a platform-optimized cancellation exception.
+ * This allows us to configure exceptions separately on JVM and other platforms.
+ */
+internal expect abstract class PlatformOptimizedCancellationException(
+    message: String? = null
+) : CancellationException
