+++ compose/foundation/foundation/src/jsNativeMain/kotlin/androidx/compose/foundation/ActualJsNative.jsNative.kt
@@ -1,7 +1,5 @@
-// ktlint-disable filename
-
 /*
- * Copyright 2021 The Android Open Source Project
+ * Copyright 2023 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,20 +16,8 @@
 
 package androidx.compose.foundation
 
-import kotlinx.coroutines.CancellationException
-
-internal actual typealias AtomicReference<V> = java.util.concurrent.atomic.AtomicReference<V>
-
-internal actual typealias AtomicLong = java.util.concurrent.atomic.AtomicLong
+import kotlin.coroutines.cancellation.CancellationException
 
 internal actual abstract class PlatformOptimizedCancellationException actual constructor(
     message: String?
-) : CancellationException(message) {
-
-    override fun fillInStackTrace(): Throwable {
-        // Avoid null.clone() on Android <= 6.0 when accessing stackTrace
-        stackTrace = emptyArray()
-        return this
-    }
-
-}
+) : CancellationException(message)
