+++ compose/ui/ui/src/commonMain/kotlin/androidx/compose/ui/Expect.kt
@@ -18,6 +18,7 @@ package androidx.compose.ui
 
 import androidx.compose.ui.node.ModifierNodeElement
 import androidx.compose.ui.platform.InspectorInfo
+import kotlinx.coroutines.CancellationException
 
 internal expect fun areObjectsOfSameType(a: Any, b: Any): Boolean
 
@@ -37,3 +38,11 @@ internal expect fun areObjectsOfSameType(a: Any, b: Any): Boolean
 internal expect fun InspectorInfo.tryPopulateReflectively(
     element: ModifierNodeElement<*>
 )
+
+/**
+ * Represents a platform-optimized cancellation exception.
+ * This allows us to configure exceptions separately on JVM and other platforms.
+ */
+internal expect abstract class PlatformOptimizedCancellationException(
+    message: String? = null
+) : CancellationException
