+++ compose/ui/ui/src/jvmMain/kotlin/androidx/compose/ui/Actual.kt
@@ -19,6 +19,7 @@ package androidx.compose.ui
 import androidx.compose.ui.node.ModifierNodeElement
 import androidx.compose.ui.platform.InspectorInfo
 import androidx.compose.ui.util.fastForEach
+import kotlinx.coroutines.CancellationException
 
 internal actual fun areObjectsOfSameType(a: Any, b: Any): Boolean {
     return a::class.java === b::class.java
@@ -46,4 +47,16 @@ internal actual fun InspectorInfo.tryPopulateReflectively(
                 }
             }
         }
+}
+
+internal actual abstract class PlatformOptimizedCancellationException actual constructor(
+    message: String?
+) : CancellationException(message) {
+
+    override fun fillInStackTrace(): Throwable {
+        // Avoid null.clone() on Android <= 6.0 when accessing stackTrace
+        stackTrace = emptyArray()
+        return this
+    }
+
 }
